move documentation to make it easier to read
